descr: setup_descrs lifecycle + dense all_descrs
Remove global descriptor registry. Add get_descr_index() /
set_descr_index() to Descr trait with AtomicI32 backing
(history.py:95, BackendDescr.descr_index = -1).

Optimizer.ensure_descr_index() assigns sequential indices
inline during serialization (descr.py:28 parity), breaking
the prior circular dependency. all_descrs threads through
MIStaticData ↔ Optimizer via std::mem::take.

all_descrs is now Vec<DescrRef> (dense, not sparse Option).
bridgeopt uses direct all_descrs[idx] indexing (bridgeopt.py:155).
heap.rs export filters get_descr_index() == -1 (heap.py:828).
assert len < 2^15 at registration (descr.py:47).

+++ majit/majit-metainterp/src/resume.rs
@@ -3688,7 +3688,7 @@ impl ResumeDataLoopMemo {
                         &numb_state.liveboxes,
                         &new_liveboxes,
                     ) as i32);
-                    numb_state.append_int(descr_idx as i32);
+                    numb_state.append_int(descr_idx);
                     numb_state.writer.append_short(self._gettagged(
                         val,
                         env,
@@ -3715,7 +3715,7 @@ impl ResumeDataLoopMemo {
                         &new_liveboxes,
                     ) as i32);
                     numb_state.append_int(index as i32);
-                    numb_state.append_int(descr_idx as i32);
+                    numb_state.append_int(descr_idx);
                     numb_state.writer.append_short(self._gettagged(
                         val,
                         env,
@@ -3996,10 +3996,10 @@ pub struct TaggedPendingField {
 pub struct OptimizerKnowledgeForResume {
     /// (obj_opref, descr_index, val_opref) heap field triples.
     /// bridgeopt.py:96-101
-    pub heap_fields: Vec<(majit_ir::OpRef, u32, majit_ir::OpRef)>,
+    pub heap_fields: Vec<(majit_ir::OpRef, i32, majit_ir::OpRef)>,
     /// (array_opref, index, descr_index, val_opref) heap array item quads.
     /// bridgeopt.py:102-108
-    pub heap_arrayitems: Vec<(majit_ir::OpRef, i64, u32, majit_ir::OpRef)>,
+    pub heap_arrayitems: Vec<(majit_ir::OpRef, i64, i32, majit_ir::OpRef)>,
     /// (const_func_ptr, result_opref) loop-invariant call results.
     pub loopinvariant_results: Vec<(i64, majit_ir::OpRef)>,
 }
